Move register function to tagging.registry submodule, to avoid weird architecture of the package and allow more flexibility
